web/requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.2.1
Deprecated==1.2.18
flash==1.0.3
Flask==3.1.2
Flask-Limiter==3.12
Flask-MySQL==1.6.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
mysql-connector-python==9.4.0
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.2
PyMySQL==1.1.2
rich==13.9.4
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
wrapt==1.17.3

